+++ src/questions/tableQuestions.ts
@@ -103,15 +103,30 @@ export const DefaultTableQuestion = TableQuestionSchema.parse({
   columns: [DefaultTableColumn]
 });
 
+// The available columns for the research output table question type (using the commonStandardId)
+export const ResearchOutputTableColumnsEnum = z.enum([
+  'title',
+  'description',
+  'type',
+  'data_flags',
+  'data_access',
+  'issued',
+  'byte_size',
+  'host',
+  'metadata',
+  'license_ref',
+  'custom'
+]);
+
 const ResearchOutputTableColumnPreferenceSchema = z.object({
   label: z.string().default(''),    // The label of the preference option
   value: z.string().default('')     // The value of the preference option
 });
 
-const ResearchOutputTitleColumnSchema = z.object({
+export const ResearchOutputTitleColumnSchema = z.object({
   ...TableColumnSchema.shape,
   heading: z.string().default('Title'),
-  commonStandardId: z.string().optional(),
+  commonStandardId: z.literal('title'),
   help: z.string().default('Enter the title of this research output'),
   required: z.boolean().default(true),
   content: TextQuestionSchema
@@ -122,7 +137,7 @@ const DefaultResearchOutputTitleContent = {
   label: '',
   help: '',
 }
-const DefaultResearchOutputTitleColumn = ResearchOutputTitleColumnSchema.parse({
+export const DefaultResearchOutputTitleColumn = ResearchOutputTitleColumnSchema.parse({
   // This commonStandardId is tied to how we render the dataset in the RDA Common Standard.
   // Any change will also need to be made to `buildDataset` function of
   // `src/lambda/layer/dmp.ts` in the `dmptool-infrastructure` repo.`
@@ -132,10 +147,10 @@ const DefaultResearchOutputTitleColumn = ResearchOutputTitleColumnSchema.parse({
   content: DefaultResearchOutputTitleContent
 });
 
-const ResearchOutputDescriptionColumnSchema = z.object({
+export const ResearchOutputDescriptionColumnSchema = z.object({
   ...TableColumnSchema.shape,
   heading: z.string().default('Description'),
-  commonStandardId: z.string().optional(),
+  commonStandardId: z.literal('description'),
   help: z.string().default('Enter a brief description of this research output'),
   content: TextAreaQuestionSchema
 });
@@ -145,7 +160,7 @@ const DefaultResearchOutputDescriptionContent = {
   label: '',
   help: '',
 }
-const DefaultResearchOutputDescriptionColumn = ResearchOutputDescriptionColumnSchema.parse({
+export const DefaultResearchOutputDescriptionColumn = ResearchOutputDescriptionColumnSchema.parse({
   // This commonStandardId is tied to how we render the dataset in the RDA Common Standard.
   // Any change will also need to be made to `buildDataset` function of
   // `src/lambda/layer/dmp.ts` in the `dmptool-infrastructure` repo.`
@@ -155,7 +170,7 @@ const DefaultResearchOutputDescriptionColumn = ResearchOutputDescriptionColumnSc
   content: DefaultResearchOutputDescriptionContent
 });
 
-const DefaultResearchOutputTypeOptions = [
+export const DefaultResearchOutputTypeOptions = [
   { label: 'Dataset', value: 'dataset' },
   { label: 'Software', value: 'software' },
   { label: 'Other', value: 'other' },
@@ -166,15 +181,15 @@ const DefaultResearchOutputTypeContent = SelectBoxQuestionSchema.parse({
   meta: DefaultMeta,
   options: DefaultResearchOutputTypeOptions
 });
-const ResearchOutputTypeColumnSchema = z.object({
+export const ResearchOutputTypeColumnSchema = z.object({
   ...TableColumnSchema.shape,
   heading: z.string().default('Type'),
-  commonStandardId: z.string().optional(),
+  commonStandardId: z.literal('type'),
   help: z.string().default('Select the type of this research output'),
   required: z.boolean().default(true),
   content: SelectBoxQuestionSchema,
 });
-const DefaultResearchOutputTypeColumn = ResearchOutputTypeColumnSchema.parse({
+export const DefaultResearchOutputTypeColumn = ResearchOutputTypeColumnSchema.parse({
   // This commonStandardId is tied to how we render the dataset in the RDA Common Standard.
   // Any change will also need to be made to `buildDataset` function of
   // `src/lambda/layer/dmp.ts` in the `dmptool-infrastructure` repo.`
@@ -184,7 +199,7 @@ const DefaultResearchOutputTypeColumn = ResearchOutputTypeColumnSchema.parse({
   content: DefaultResearchOutputTypeContent,
 });
 
-const DefaultResearchOutputDataFlagsOptions = [
+export const DefaultResearchOutputDataFlagsOptions = [
   { label: 'May contain sensitive data?', value: 'sensitive', checked: false },
   { label: 'May contain personally identifiable information?', value: 'personal', checked: false },
 ];
@@ -194,15 +209,15 @@ const DefaultResearchOutputDataFlagsContent = CheckboxesQuestionSchema.parse({
   meta: DefaultMeta,
   options: DefaultResearchOutputDataFlagsOptions
 });
-const ResearchOutputDataFlagsColumnSchema = z.object({
+export const ResearchOutputDataFlagsColumnSchema = z.object({
   ...TableColumnSchema.shape,
   heading: z.string().default('Data Flags'),
-  commonStandardId: z.string().optional(),
+  commonStandardId: z.literal('data_flags'),
   help: z.string().default('Mark all of the statements that are true about the dataset'),
   enabled: z.boolean().default(false),
   content: CheckboxesQuestionSchema,
 })
-const DefaultResearchOutputDataFlagsColumn = ResearchOutputDataFlagsColumnSchema.parse({
+export const DefaultResearchOutputDataFlagsColumn = ResearchOutputDataFlagsColumnSchema.parse({
   // This commonStandardId is tied to how we render the dataset in the RDA Common Standard.
   // Any change will also need to be made to `buildDataset` function of
   // `src/lambda/layer/dmp.ts` in the `dmptool-infrastructure` repo.`
@@ -212,7 +227,7 @@ const DefaultResearchOutputDataFlagsColumn = ResearchOutputDataFlagsColumnSchema
   content: DefaultResearchOutputDataFlagsContent
 });
 
-const DefaultResearchOutputAccessLevelOptions = [
+export const DefaultResearchOutputAccessLevelOptions = [
   { label: 'Open Access', value: 'open' },
   { label: 'Restricted Access', value: 'restricted' },
   { label: 'Other', value: 'closed' },
@@ -223,10 +238,10 @@ const DefaultResearchOutputAccessLevelContent = RadioButtonsQuestionSchema.parse
   meta: DefaultMeta,
   options: DefaultResearchOutputAccessLevelOptions
 });
-const ResearchOutputAccessLevelColumnSchema = z.object({
+export const ResearchOutputAccessLevelColumnSchema = z.object({
   ...TableColumnSchema.shape,
   heading: z.string().default('Access Level'),
-  commonStandardId: z.string().optional(),
+  commonStandardId: z.literal('data_access'),
   help: z.string().default('Select the access level for this research output'),
   enabled: z.boolean().default(false),
   content: RadioButtonsQuestionSchema,
@@ -241,15 +256,15 @@ export const DefaultResearchOutputAccessLevelColumn = ResearchOutputAccessLevelC
   content: DefaultResearchOutputAccessLevelContent
 });
 
-const ResearchOutputReleaseDateColumnSchema = z.object({
+export const ResearchOutputReleaseDateColumnSchema = z.object({
   ...TableColumnSchema.shape,
   heading: z.string().default('Anticipated Release Date'),
-  commonStandardId: z.string().optional(),
+  commonStandardId: z.literal('issued'),
   help: z.string().default('The anticipated release date for the research output'),
   enabled: z.boolean().default(false),
   content: DateQuestionSchema,
 })
-const DefaultResearchOutputReleaseDateColumn = ResearchOutputReleaseDateColumnSchema.parse({
+export const DefaultResearchOutputReleaseDateColumn = ResearchOutputReleaseDateColumnSchema.parse({
   // This commonStandardId is tied to how we render the dataset in the RDA Common Standard.
   // Any change will also need to be made to `buildDataset` function of
   // `src/lambda/layer/dmp.ts` in the `dmptool-infrastructure` repo.`
@@ -259,7 +274,7 @@ const DefaultResearchOutputReleaseDateColumn = ResearchOutputReleaseDateColumnSc
   content: DefaultDateQuestion,
 });
 
-const DefaultResearchOutputByteSizeOptions = [
+export const DefaultResearchOutputByteSizeOptions = [
   { label: 'bytes', value: 'bytes' },
   { label: 'KB (kilobytes)', value: 'kb' },
   { label: 'MB (megabytes)', value: 'mb' },
@@ -276,15 +291,15 @@ const DefaultResearchOutputByteSizeContent = NumberWithContextQuestionSchema.par
   attributes: DefaultResearchOutputByteSizeContentAttributes,
   meta: DefaultMeta,
 });
-const ResearchOutputByteSizeColumnSchema = z.object({
+export const ResearchOutputByteSizeColumnSchema = z.object({
   ...TableColumnSchema.shape,
   heading: z.string().default('Byte Size'),
-  commonStandardId: z.string().optional(),
+  commonStandardId: z.literal('byte_size'),
   help: z.string().default('The size of the research output in bytes'),
   enabled: z.boolean().default(false),
   content: NumberWithContextQuestionSchema,
 });
-const DefaultResearchOutputByteSizeColumn = ResearchOutputByteSizeColumnSchema.parse({
+export const DefaultResearchOutputByteSizeColumn = ResearchOutputByteSizeColumnSchema.parse({
   // This commonStandardId is tied to how we render the dataset in the RDA Common Standard.
   // Any change will also need to be made to `buildDataset` function of
   // `src/lambda/layer/dmp.ts` in the `dmptool-infrastructure` repo.`
@@ -294,16 +309,16 @@ const DefaultResearchOutputByteSizeColumn = ResearchOutputByteSizeColumnSchema.p
   content: DefaultResearchOutputByteSizeContent
 });
 
-const ResearchOutputRepositoryColumnSchema = z.object({
+export const ResearchOutputRepositoryColumnSchema = z.object({
   ...TableColumnSchema.shape,
   heading: z.string().default('Repository'),
-  commonStandardId: z.string().optional(),
+  commonStandardId: z.literal('host'),
   help: z.string().default('Select repository(ies) you would prefer users to deposit in'),
   enabled: z.boolean().default(false),
   content: RepositorySearchQuestionSchema,
   preferences: z.array(ResearchOutputTableColumnPreferenceSchema).default([])
 });
-const DefaultResearchOutputRepositoryColumn = ResearchOutputRepositoryColumnSchema.parse({
+export const DefaultResearchOutputRepositoryColumn = ResearchOutputRepositoryColumnSchema.parse({
   heading: 'Repository(ies)',
   commonStandardId: 'host',
   enabled: false,
@@ -311,16 +326,16 @@ const DefaultResearchOutputRepositoryColumn = ResearchOutputRepositoryColumnSche
   preferences: []
 });
 
-const ResearchOutputMetadataStandardColumnSchema = z.object({
+export const ResearchOutputMetadataStandardColumnSchema = z.object({
   ...TableColumnSchema.shape,
   heading: z.string().default('Metadata Standard'),
-  commonStandardId: z.string().optional(),
+  commonStandardId: z.literal('metadata'),
   help: z.string().default('Select metadata standard(s) you would prefer users to use'),
   enabled: z.boolean().default(false),
   content: MetadataStandardSearchQuestionSchema,
   preferences: z.array(ResearchOutputTableColumnPreferenceSchema).default([])
 });
-const DefaultResearchOutputMetadataStandardColumn = ResearchOutputMetadataStandardColumnSchema.parse({
+export const DefaultResearchOutputMetadataStandardColumn = ResearchOutputMetadataStandardColumnSchema.parse({
   // This commonStandardId is tied to how we render the dataset in the RDA Common Standard.
   // Any change will also need to be made to `buildDataset` function of
   // `src/lambda/layer/dmp.ts` in the `dmptool-infrastructure` repo.`
@@ -331,16 +346,16 @@ const DefaultResearchOutputMetadataStandardColumn = ResearchOutputMetadataStanda
   preferences: []
 });
 
-const ResearchOutputLicenseColumnSchema = z.object({
+export const ResearchOutputLicenseColumnSchema = z.object({
   ...TableColumnSchema.shape,
   heading: z.string().default('License'),
-  commonStandardId: z.string().optional(),
+  commonStandardId: z.literal('license_ref'),
   help: z.string().default('Select the license you will apply to the research output'),
   enabled: z.boolean().default(false),
   content: LicenseSearchQuestionSchema,
   preferences: z.array(ResearchOutputTableColumnPreferenceSchema).default([])
 });
-const DefaultResearchOutputLicenseColumn = ResearchOutputLicenseColumnSchema.parse({
+export const DefaultResearchOutputLicenseColumn = ResearchOutputLicenseColumnSchema.parse({
   // This commonStandardId is tied to how we render the dataset in the RDA Common Standard.
   // Any change will also need to be made to `buildDataset` function of
   // `src/lambda/layer/dmp.ts` in the `dmptool-infrastructure` repo.`
@@ -372,18 +387,20 @@ const DefaultResearchOutputCustomContent = ResearchOutputCustomContentSchema.par
 export const ResearchOutputCustomColumnSchema = z.object({
   ...TableColumnSchema.shape,
   heading: z.string().default('Custom Column'),
+  commonStandardId: z.literal('custom'),
   help: z.string().default('Explanation of what we expect the user to enter.'),
   enabled: z.boolean().default(false),
   content: ResearchOutputCustomContentSchema,
 });
 export const DefaultResearchOutputCustomColumn = ResearchOutputCustomColumnSchema.parse({
   heading: 'Custom Column',
+  commonStandardId: 'custom',
   enabled: false,
   content: DefaultResearchOutputCustomContent
 });
 
 // Add this BEFORE ResearchOutputTableQuestionSchema
-const AnyResearchOutputColumnSchema = z.union([
+export const AnyResearchOutputColumnSchema = z.union([
   ResearchOutputTitleColumnSchema,
   ResearchOutputDescriptionColumnSchema,
   ResearchOutputTypeColumnSchema,
@@ -425,6 +442,16 @@ export const DefaultResearchOutputTableQuestion = ResearchOutputTableQuestionSch
 // This will ensure that object validations are against the Zod schemas defined above
 export type TableQuestionType = z.infer<typeof TableQuestionSchema>;
 export type AnyTableColumnQuestionType = z.infer<typeof AnyTableColumnQuestionSchema>;
+export type ResearchOutputTitleColumnType = z.infer<typeof ResearchOutputTitleColumnSchema>;
+export type ResearchOutputDescriptionColumnType = z.infer<typeof ResearchOutputDescriptionColumnSchema>;
+export type ResearchOutputTypeColumnType = z.infer<typeof ResearchOutputTypeColumnSchema>;
+export type ResearchOutputDataFlagsColumnType = z.infer<typeof ResearchOutputDataFlagsColumnSchema>;
+export type ResearchOutputAccessLevelColumnType = z.infer<typeof ResearchOutputAccessLevelColumnSchema>;
+export type ResearchOutputReleaseDateColumnType = z.infer<typeof ResearchOutputReleaseDateColumnSchema>;
+export type ResearchOutputByteSizeColumnType = z.infer<typeof ResearchOutputByteSizeColumnSchema>;
+export type ResearchOutputRepositoryColumnType = z.infer<typeof ResearchOutputRepositoryColumnSchema>;
+export type ResearchOutputMetadataStandardColumnType = z.infer<typeof ResearchOutputMetadataStandardColumnSchema>;
+export type ResearchOutputLicenseColumnType = z.infer<typeof ResearchOutputLicenseColumnSchema>;
 export type ResearchOutputCustomTableColumnType = z.infer<typeof ResearchOutputCustomColumnSchema>;
 export type AnyResearchOutputColumnType = z.infer<typeof AnyResearchOutputColumnSchema>;
 export type ResearchOutputTableQuestionType = z.infer<typeof ResearchOutputTableQuestionSchema>;