+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # dmptool-types CHANGELOG
 
+## v3.2.0
+- Added new exports for `ResearchOutputTable` columns and mappings for commonStandardId to Types, Schemas and Defaults
+- Removed `optional` flag from the `commonStandardId` property on `researchOutputTable` questions and answers, since we need this to map entries from the `dataset` array in the RDA common standard 
+- Updated dev dependencies
+
 ## v3.1.7
 - Made `id` properties in `narrative` portion of the DMP Tool extensions optional